AI_Voice/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.1

# AWS Bedrock
boto3==1.35.0
botocore==1.35.0

# Deepgram
deepgram-sdk==3.10.1

# Database (asyncpg is all we need for now - pure Python, no PostgreSQL required)
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
# psycopg2-binary==2.9.9  # Not needed for MVP - asyncpg handles PostgreSQL

# Redis
redis==5.2.1
hiredis==3.0.0

# Environment & Config
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.0

# HTTP Client
httpx==0.28.1
aiohttp==3.11.11

# Audio Processing
numpy==2.2.1
# pyaudio==0.2.14  # Not needed for web-based audio (browser handles it)
audioop-lts==0.2.1  # For mu-law conversion (phone support)

# Utilities
python-json-logger==3.2.1
prometheus-client==0.21.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.14.1
